feat(treesit): basic treesit configs with `treesit-auto'
Emacs 29.1 has shipped with the treesit package. The command
`treesit-install-language-grammar' find recipes from the variable
`treesit-language-source-alist', which is originally nil. The convenient package
treesit-auto maintains a list of recipes and appends rules to
`major-mode-remap-alist', tremendously easing the trouble to configure treesit
in Emacs.

I prefer to install *all* tree-sitter parsers manually, so the behavior is
predictable (and I do not want to enable treesit *everywhere* since it may be
unstable). So I write me a function `+treesit-auto-install' to help with the
stuff.