feat(ui): better font settings
By applying different fonts in different major-mode, a writer can easily
distinguish the current context, which eases the efforts of context-switch and
makes the UI never be so boring.

To implement this, I use three customizable variables: default, other, CJK. The
last one should be used whenever a CJK font is displayed. However, it will be
pretty cool if fonts vary in the same buffer, for different kinds of characters.
So I patch the current fontset by running `+fontset-setup-cjk' function, which
applies `celeste-cjk-font-name' to all CJK (and some Unicode symbols)
characters. To change the fontset for different buffers, I leverage
`buffer-face-mode', which makes buffer-local fonts possible.