feat(evil): goodbye, evil-mode!
There are several reasons why I abandon evil-mode after two-year Vim usage:

1. Despite its age and maturity, evil-mode is still a complicated package. The
codebase is too overwhelming to understand, with hacks and magics almost
everywhere. It's difficult to figure out what is causing the problem when it
actually happens.

2. Evil-mode does the right thing at most of the time, but evil-collection does
not. The latter is the complement for vim-like keybindings in other modes. But
evil-collection changes rapidly, and its collected bindings are too aggressive,
which makes my customization a cumbersome hell.

3. Keybinding override is an annoying problem. Evil-mode and evil-collection
break default bindings of many major modes. Just being primitive is the best
solution to this problem.

4. Evil consumes a lot of startup time. Though byte-compilation can accelerate a
little, it's still nonnegligible.

5. How about other Vim features? Frankly speaking, I like the idea of
"textobject". But nothing is a patent of Vim! In Emacs, with other modal editing
mode like meow, I can still enjoy the efficiency of modal editing!