fix(tempel): modify `tempel-trigger-prefix`.
After merge of this patch https://github.com/manateelazycat/lsp-bridge/pull/917,
there are still some problems about `tempel-trigger-prefix`. For example, in
org-mode, if the `tempel-trigger-prefix` is something like `<t<`, tempel
completion won't be matched at all.

I've found some reasons. It seems that, in org-mode (and some other modes), the
angle brackets are not recognized as the keyword for matching completions. Maybe
it is designed to be like this. I recommend to not use a prefix containing some
special symbols. `t|` is my choice now.