feat(project): remove projectile and perspective
That was a lonely night, when I was cold and exhausted, staring my screen,
feeling hopeless.

I've been tired of the destructive bugs of perspective.el. It sometimes broke my
workspace, sometimes even prevented me from killing any buffer because of its
harmful hooks. I had to `save-buffers-kill-emacs`.

No much fortune when it came to projectile. It was not as OOTB as it claimed it
to be. It could not recognize my project even though a `.projectile` is put into
the root path. It could not ignore files/directories specified in `.projectile`.
And its functionalities had been copied to project.el...

I also tried tabspaces, which reuses builtin project.el and tab-bar.el. But man,
you could see, it did not work as I expected in many cases...

I began to ask me, was it necessary to use a so-called "workspace manager"?
Maybe not, maybe yes. What I know certainly is that, I do wish to control my
work flow. Therefore, even a package provides a feature that is really
attractive to me, I shouldn't pick it if it behaves unpredictably. Either too
buggy, or undocumented. That's not the life I wish to be.

And, after all, less is more.