fix(mini): whether open *scratch* on startup
Do not set the initial buffer to *scratch* if Emacs is invoked in TUI mode.
This is because that if Emacs is invoked in TUI mode (with option `-nw`),
probably it's used as the default editor ($EDITOR), therefore a filename is
usually specified. I want to see the buffer of the file immediately in this
case, not the *scratch* buffer.