sis: remove
Ok, sis (smart-input-source) is finally deprecated.

I think it's kind of useful, but it's really buggy and poorly-designed. It makes
C-g acts strangely and slows down input stream because of heavy usage of
`post-command-hook`, and its functionality sometimes fails out. Also It's hard
to debug. Each time something went wrong with sis, I have no choice but restart
my Emacs.

Oh, and I tried to type Chinese without sis. It's natural and not so bad. I
think I imported this package because I used to be an Evil user. Now Evil is
deprecated in my configuration, maybe it's also the time to remove sis. Goodbye!