lang: use `celeste/setup-lang` instead of `ide-zero-define`
`celeste/setup-lang` is the rewrite of `ide-zero-define` with less premature
design and focus on satisfying my own use cases. `celeste/setup-lang` does not
import complex abstractions to conceal the backends under the hood, but targets
the specified packages like Eglot and Flycheck directly. The macro is also
rewritten to be more standard and byte-compilation friendly.