gptel: adapt for new version
Fix the problem when interacting with newer transient.el. This is actually the
problem of ace-window, I guess, but I just put the fixup here at this moment.

Set `gptel-expert-commands`, because I want to see more options such as number
of messages to send.

Set `gptel--num-messages-to-send` to zero by default. This saves a lot of cost.